Dashboard/requirements.txt
echo "dash>=2.14.0
dash-bootstrap-components>=1.5.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
plotly>=5.15.0
openpyxl>=3.1.0
scikit-learn>=1.3.0" > requirements.txt